backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.1.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
apscheduler==3.10.4
pytest==7.4.0
httpx==0.25.0
slowapi==0.1.9